Improved {set,bag} hash functions
Now the functions take into account hash functions for elements.

It is also necessary to walk over the whole hash table because
the iteration order of hash-table-walk is unspecified. We need
to make sure that equal qmaps have equal hashes, even if their
elements are in different order in hash tables.

Also added more tests