Version 1.0.0 release
This is a portable implementation of the full API of SRFI 113.
It is integrated with SRFI 114 default comparator (which requires
specific SRFI 114 implementation, unfortunately). The backend of
implementation is abstracted out into so-called qmaps. This provides
means for a more efficient implementation in case SRFI 69 hash tables
are too slow.